Added Database, Logger, mysql, database use, etc
I have added a Database class which provides a connection to the database. I also created the mysql.sql file for populating the database with tables.

I then added support to the various user files for modifying the database entries. Only the UserBase and UserTemp have been fully implemented for accessing the database so far.

To remove the need for extra data in basic user entries and classes, I decided to make admins seperate, with their own UserAdmin class and table in the database.

I also began to implement actual web pages.